add python-hcl2 dependency
