+++ tuvok/__version__.py
@@ -1,0 +1,5 @@
+__title__ = 'tuvok'
+__version__ = '0.1.0'
+__license__ = 'Apache 2.0'
+__copyright__ = 'Copyright Rackspace US, Inc. 2018'
+
